+++ app/routes/components/Switch/index.tsx
@@ -1,15 +1,14 @@
-import type { Route } from "./+types/index";
-
 import { ComponentTemplate } from "src/components/ComponentTemplate";
 import { Section } from "src/components/Section";
 
-import { sourceCode } from "./source_code.codegen";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import type { Route } from "./+types/index";
+
 import { changelog } from "./changelog.codegen";
-
 import { Demo } from "./Demo";
+import { demoSourceCode } from "./Demo_source_code.codegen";
+import { sourceCode } from "./source_code.codegen";
 
-export function meta({}: Route.MetaArgs) {
+export function meta(_: Route.MetaArgs) {
   return [
     { title: "Switch | Copy UI" },
     { name: "description", content: "The Switch component from Copy UI." },