+++ src/components/Switch/Switch.tsx
@@ -1,54 +1,81 @@
-import { type CSSProperties, type FC } from "react";
 import { merge } from "es-toolkit";
+import type { CSSProperties, FC } from "react";
 
 import { useTheme } from "../ThemeProvider";
 
 import styles from "./Switch.module.scss";
 
 export type SwitchProps = {
-  size?: 'sm' | 'md' | 'lg';
+  size?: "text" | "xs" | "sm" | "md" | "lg" | "xl";
   value: boolean;
   onChange?: (value: boolean) => void;
 };
 
 export const Switch: FC<SwitchProps> = (props) => {
-  const { size = 'md', value, onChange } = props;
+  const { size = "md", value, onChange } = props;
 
   const theme = useTheme();
 
   const computedStyle = mergeStyles([
-    size === 'sm' && {
-      '--switch-rail-width': '3.5rem',
-      '--switch-rail-height': theme.tokens.inputBaseSmHeight,
-    } as CSSProperties,
-    size === 'md' && {
-      '--switch-rail-width': '4rem',
-      '--switch-rail-height': theme.tokens.inputBaseMdHeight,
-    } as CSSProperties,
-    size === 'lg' && {
-      '--switch-rail-width': '4.5rem',
-      '--switch-rail-height': theme.tokens.inputBaseLgHeight,
-    } as CSSProperties,
+    size === "text" &&
+      ({
+        "--switch-rail-width": "1.75rem",
+        "--switch-rail-height": "1rem",
+        "--switch-thumb-gap": "2.5px",
+      } as CSSProperties),
+    size === "xs" &&
+      ({
+        "--switch-rail-width": "2.5rem",
+        "--switch-rail-height": `calc(${theme.tokens.inputBaseXsHeight} * 0.8)`,
+        "--switch-thumb-gap": "2.5px",
+      } as CSSProperties),
+    size === "sm" &&
+      ({
+        "--switch-rail-width": "2.875rem",
+        "--switch-rail-height": `calc(${theme.tokens.inputBaseSmHeight} * 0.8)`,
+        "--switch-thumb-gap": "3px",
+      } as CSSProperties),
+    size === "md" &&
+      ({
+        "--switch-rail-width": "3.25rem",
+        "--switch-rail-height": `calc(${theme.tokens.inputBaseMdHeight} * 0.8)`,
+        "--switch-thumb-gap": "3px",
+      } as CSSProperties),
+    size === "lg" &&
+      ({
+        "--switch-rail-width": "3.625rem",
+        "--switch-rail-height": `calc(${theme.tokens.inputBaseLgHeight} * 0.8)`,
+        "--switch-thumb-gap": "3.5px",
+      } as CSSProperties),
+    size === "xl" &&
+      ({
+        "--switch-rail-width": "4.25rem",
+        "--switch-rail-height": `calc(${theme.tokens.inputBaseXlHeight} * 0.8)`,
+        "--switch-thumb-gap": "3.5px",
+      } as CSSProperties),
     {
-      '--switch-rail-bg': theme.colors.gray['300'],
-      '--switch-rail-checked-bg': theme.colors.blue['600'],
-      '--switch-thumb-bg': 'white',
+      "--switch-rail-bg": theme.colors.gray["300"],
+      "--switch-rail-checked-bg": theme.colors.blue["600"],
+      "--switch-thumb-bg": "white",
     } as CSSProperties,
   ]);
 
   return (
-    <div
+    <button
       style={computedStyle}
       className={styles.rail}
       data-checked={value}
       onClick={() => onChange?.(!value)}
+      type="button"
+      role="switch"
+      aria-checked={value}
     >
       <div className={styles.thumb} />
-    </div>
+    </button>
   );
 };
 
-Switch.displayName = 'Switch';
+Switch.displayName = "Switch";
 
 function mergeStyles(styles: (CSSProperties | false | undefined)[]) {
   return styles.reduce((prev, next) => {